backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.1
python-jose==3.3.0
python-multipart==0.0.6
pypdf2==3.0.1
pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
google-generativeai==0.3.1
google-api-python-client==2.108.0
